refacto some params and change directory for locale
