+++ src/app/[locale]/globals.scss
