+++ src/app/[locale]/layout.tsx
