+++ src/app/[locale]/movies/(search)/SearchResults.module.scss
