+++ src/app/[locale]/movies/(search)/genres/[id]/page.tsx
