+++ src/app/[locale]/movies/(search)/layout.module.scss
