+++ src/app/[locale]/movies/(search)/layout.tsx
