+++ src/app/[locale]/movies/[id]/page.tsx
@@ -1,6 +1,7 @@
 import MovieDetails from '@/components/movie-details/MovieDetails';
 import SimilarMovies from '@/components/similar-movies/SimilarMovies';
 import { IMovieDetail } from '@/interfaces/movie';
+import { IParams } from '@/interfaces/params';
 import { getMovieByPath } from '@/utils/movieClient';
 import { notFound } from 'next/navigation';
 import React, { Suspense } from 'react';
@@ -8,8 +9,8 @@ import React, { Suspense } from 'react';
 export const dynamic = "force-static";
 export const revalidate = 3600;
 
-const MoviesDetail = async ({ params }: { params: { id: string } }) => {
-    const movie: IMovieDetail = await getMovieByPath(`/movie/${params.id}`);
+const MoviesDetail = async ({ params: { id, locale } }: IParams) => {
+    const movie: IMovieDetail = await getMovieByPath(`/movie/${id}`, [], locale);
 
     if (!movie.original_title) {
         return notFound();
