+++ src/interfaces/params.ts
@@ -1,0 +1,15 @@
+export interface IParams {
+  params: {
+    id?: string;
+    locale?: string;
+  };
+  key?: string;
+  value?: string;
+}
+
+export interface ISearchParams {
+  sort_by: string;
+  release_gte: string;
+  release_lte: string;
+  whith_genres?: string;
+}
